+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   settings as the main video stream.
 * Always copy audio unless `--acodec` or `--downmix-to-stereo` are specified. Previously would
   re-encode to opus when changing container.
+* Add `--encoder` support for qsv family of ffmpeg encoders: av1_qsv, hevc_qsv, vp9_qsv, h264_qsv and mpeg2_qsv.
+* Enable default look_ahead=1, lookahead-depth=40 args for encoders: av1_qsv, hevc_qsv, h264_qsv.
 
 # v0.7.2
 * Print failing ffmpeg stderr output.